+++ Plugins/UI/Source/NexusUI/Private/Widgets/NDeveloperOverlay.cpp
@@ -52,7 +52,7 @@ void UNDeveloperOverlay::ShowContainerBanner(const FText& Text, ENColor MessageC
 	{
 		ContainerBanner->SetVisibility(ESlateVisibility::SelfHitTestInvisible);
 		ContainerBanner->SetBrushColor(FNColor::GetLinearColor(BannerColor));
-		if (!Text.IsEmpty())
+		if (!Text.IsEmpty() && IsValid(ContainerBannerMessage))
 		{
 			ContainerBannerMessage->SetText(Text);
 			ContainerBannerMessage->SetColorAndOpacity(FNColor::GetLinearColor(MessageColor));